# docs/swagger.yaml
basePath: /api/v2/order
...
host: localhost:5000
info:
  contact: {}
  description: This is a server for Latipe Order Service.
  title: API Documentation
  version: "2.0"
paths:
...
  /admin/statistic/total-commission-order-in-day:
    get:
      consumes:
      - application/json
      description: Get total commission order in day
      parameters:
      - description: date
        in: query
        name: date
        type: string
      produces:
      - application/json
      responses: {}
      summary: Get total commission order in day
      tags:
      - Statistic
  /admin/statistic/total-commission-order-in-month:
    get:
      consumes:
      - application/json
      description: Get total commission order in month
      parameters:
      - description: date
        in: query
        name: date
        type: string
      produces:
      - application/json
      responses: {}
      summary: Get total commission order in month
      tags:
      - Statistic
  /admin/statistic/total-order-in-day:
    get:
      consumes:
      - application/json
      description: Get total order in system in day
      parameters:
      - description: date
        in: query
        name: date
        type: string
      produces:
      - application/json
      responses: {}
      summary: Get total order in system in day
      tags:
      - Statistic
  /admin/statistic/total-order-in-month:
    get:
      consumes:
      - application/json
      description: Get total order in system in month
      parameters:
      - description: date
        in: query
        name: date
        type: string
      produces:
      - application/json
      responses: {}
      summary: Get total order in system in month
      tags:
      - Statistic
  /admin/statistic/total-order-in-year:
    get:
      consumes:
      - application/json
      description: Get total order in system in year
      parameters:
      - description: date
        in: query
        name: date
        type: string
      produces:
      - application/json
      responses: {}
      summary: Get total order in system in year
      tags:
      - Statistic
...
  /store/statistic/total-commission-order-in-month:
    get:
      consumes:
      - application/json
      description: Get total commission order in month of store
      parameters:
      - description: date
        in: query
        name: date
        type: string
      produces:
      - application/json
      responses: {}
      summary: Get total commission order in month of store
      tags:
      - Statistic
  /store/statistic/total-order-in-day:
    get:
      consumes:
      - application/json
      description: Get total order in day of store
      parameters:
      - description: date
        in: query
        name: date
        type: string
      produces:
      - application/json
      responses: {}
      summary: Get total order in day of store
      tags:
      - Statistic
  /store/statistic/total-order-in-month:
    get:
      consumes:
      - application/json
      description: Get total order in month of store
      parameters:
      - description: date
        in: query
        name: date
        type: string
      produces:
      - application/json
      responses: {}
      summary: Get total order in month of store
      tags:
      - Statistic
  /store/statistic/total-order-in-year:
    get:
      consumes:
      - application/json
      description: Get total order in year of store
      parameters:
      - description: date
        in: query
        name: date
        type: string
      produces:
      - application/json
      responses: {}
      summary: Get total order in year of store
      tags:
      - Statistic
...
schemes:
- http
swagger: "2.0"